chore: fix frontend shell baseline checks
