+++ apps/frontend/src/context/auth-context.ts
@@ -1,0 +1,9 @@
+import { createContext } from 'react';
+
+export type AuthContextType = {
+  isAuthenticated: boolean;
+  login: () => void;
+  logout: () => void;
+};
+
+export const AuthContext = createContext<AuthContextType | null>(null);
