+++ apps/frontend/src/routes/ProtectedRoute.tsx
@@ -1,6 +1,6 @@
 import { Navigate, Outlet } from 'react-router-dom';
 
-import { useAuth } from '../context/AuthContext';
+import { useAuth } from '../context/useAuth';
 
 function ProtectedRoute() {
   const { isAuthenticated } = useAuth();
